- Segment class changed as requested by the announcement on canvas.

+++ Segment.java
@@ -77,7 +77,7 @@ public class Segment implements Serializable {
 	public long getChecksum() {
 		return checksum;
 	}
-	public void setChecksum(int checksum) {
+	public void setChecksum(long checksum) {
 		this.checksum = checksum;
 	}
 }
